fix: 修复 data cell 各种图形尺寸绘制问题
