+++ packages/s2-core/src/group/grid-group.ts
@@ -44,24 +44,25 @@ export class GridGroup extends Group {
 
     this.gridGroup.clear();
 
+    const verticalBorderWidth = style.verticalBorderWidth;
+
     this.gridInfo = gridInfo;
     if (shouldDrawLeftBorder) {
-      this.gridInfo.cols.unshift(0);
+      this.gridInfo.cols.unshift(verticalBorderWidth);
     }
 
-    // line 在绘制时，包围盒计算有点问题，会代入lineWidth
+    // line 在绘制时，包围盒计算有点问题，会带入lineWidth
     // 比如传入的 x1=0, x2=10, lineWidth=20
     // 最后line得出来的包围盒 minX=-10, maxX=20，会将lineWidth/2纳入计算中
     // 最后就会导致更新过程中，GridGroup的包围盒不断被放大
     // 因此在传入时，将这部分坐标减去，并结合lineCap将这部分绘制出来，达到内容区域绘制不变，包围盒计算正确的目的
-    const verticalBorderWidth = style.verticalBorderWidth;
-    const halfVerticalBorderWidthBorderWidth = style.verticalBorderWidth / 2;
+    const halfVerticalBorderWidthBorderWidth = verticalBorderWidth / 2;
     this.gridInfo.cols.forEach((x) => {
       renderLine(
         this.gridGroup as Group,
         {
-          x1: x,
-          x2: x,
+          x1: x - halfVerticalBorderWidthBorderWidth,
+          x2: x - halfVerticalBorderWidthBorderWidth,
           y1: Math.ceil(bbox.minY + halfVerticalBorderWidthBorderWidth),
           y2: Math.floor(bbox.maxY - halfVerticalBorderWidthBorderWidth),
         },
@@ -75,15 +76,15 @@ export class GridGroup extends Group {
     });
 
     const horizontalBorderWidth = style.horizontalBorderWidth;
-    const halfHorizontalBorderWidth = style.horizontalBorderWidth / 2;
+    const halfHorizontalBorderWidth = horizontalBorderWidth / 2;
     this.gridInfo.rows.forEach((y) => {
       renderLine(
         this.gridGroup as Group,
         {
           x1: Math.ceil(bbox.minX + halfHorizontalBorderWidth),
           x2: Math.floor(bbox.maxX - halfHorizontalBorderWidth),
-          y1: y,
-          y2: y,
+          y1: y - halfHorizontalBorderWidth,
+          y2: y - halfHorizontalBorderWidth,
         },
         {
           stroke: style.horizontalBorderColor,
